requirements.txt
fastapi
uvicorn
psycopg2
sqlalchemy
google-generativeai
pypdf
python-multipart
requests
beautifulsoup4
python-dotenv
